refactor: improve app sources and rename directAPKLink to direct_apk_link
- Rename directAPKLink.dart to direct_apk_link.dart for consistency
- Update various app sources with quality improvements
- Update enhanced_app.dart model
- Update source_provider.dart
